+++ src/main/java/com/cosmian/jna/findex/FindexCloud.java
@@ -15,10 +15,47 @@ import com.sun.jna.ptr.IntByReference;
 
 public final class FindexCloud extends FindexBase {
 
+    public static String generateNewToken(
+                              String indexId,
+                              byte[] fetchEntriesSeed,
+                              byte[] fetchChainsSeed,
+                              byte[] upsertEntriesSeed,
+                              byte[] insertChainsSeed)
+        throws CloudproofException {
+
+        try (
+            final Memory fetchEntriesSeedPointer = new Memory(fetchEntriesSeed.length);
+            final Memory fetchChainsSeedPointer = new Memory(fetchChainsSeed.length);
+            final Memory upsertEntriesSeedPointer = new Memory(upsertEntriesSeed.length);
+            final Memory insertChainsSeedPointer = new Memory(insertChainsSeed.length);) {
+            fetchEntriesSeedPointer.write(0, fetchEntriesSeed, 0, fetchEntriesSeed.length);
+            fetchChainsSeedPointer.write(0, fetchChainsSeed, 0, fetchChainsSeed.length);
+            upsertEntriesSeedPointer.write(0, upsertEntriesSeed, 0, upsertEntriesSeed.length);
+            insertChainsSeedPointer.write(0, insertChainsSeed, 0, insertChainsSeed.length);
+
+            byte[] tokenBuffer = new byte[200];
+            IntByReference tokenBufferSize = new IntByReference(tokenBuffer.length);
+
+            unwrap(INSTANCE.h_generate_new_token(
+                tokenBuffer,
+                tokenBufferSize,
+                indexId,
+                fetchEntriesSeedPointer, fetchEntriesSeed.length,
+                fetchChainsSeedPointer, fetchChainsSeed.length,
+                upsertEntriesSeedPointer, upsertEntriesSeed.length,
+                insertChainsSeedPointer, insertChainsSeed.length));
+
+            byte[] tokenBytes = Arrays.copyOfRange(tokenBuffer, 0, tokenBufferSize.getValue());
+
+            return new String(tokenBytes, StandardCharsets.UTF_8);
+        }
+    }
+
     public static void upsert(
                               String token,
                               byte[] label,
-                              Map<IndexedValue, Set<Keyword>> indexedValuesAndWords,
+                              Map<IndexedValue, Set<Keyword>> additions,
+                              Map<IndexedValue, Set<Keyword>> deletions,
                               String baseUrl)
         throws CloudproofException {
 
@@ -30,52 +67,41 @@ public final class FindexCloud extends FindexBase {
             unwrap(INSTANCE.h_upsert_cloud(
                 token,
                 labelPointer, label.length,
-                indexedValuesToJson(indexedValuesAndWords),
+                indexedValuesToJson(additions),
+                indexedValuesToJson(deletions),
                 baseUrl));
         }
     }
 
     public static void upsert(IndexRequest request)
         throws CloudproofException {
-        upsert(request.token, request.label, request.indexedValuesAndWords, request.baseUrl);
+        upsert(request.token, request.label, request.additions, request.deletions, request.baseUrl);
     }
 
     public static void upsert(
                               String token,
                               byte[] label,
-                              Map<IndexedValue, Set<Keyword>> indexedValuesAndWords)
+                              Map<IndexedValue, Set<Keyword>> additions,
+                              Map<IndexedValue, Set<Keyword>> deletions)
         throws CloudproofException {
-        upsert(token, label, indexedValuesAndWords, null);
+        upsert(token, label, additions, deletions, null);
     }
 
     public static SearchResults search(SearchRequest request)
         throws CloudproofException {
-        return search(request.token, request.label, request.keywords, request.maxResultsPerKeyword, request.maxDepth,
-            request.maxDepth, request.baseUrl);
+        return search(request.token, request.label, request.keywords, request.baseUrl);
     }
 
     public static SearchResults search(String token,
                                        byte[] label,
                                        Set<Keyword> keyWords)
         throws CloudproofException {
-        return search(token, label, keyWords, 0, -1, 0, null);
+        return search(token, label, keyWords, null);
     }
 
     public static SearchResults search(String token,
                                        byte[] label,
                                        Set<Keyword> keyWords,
-                                       int maxResultsPerKeyword,
-                                       int maxDepth)
-        throws CloudproofException {
-        return search(token, label, keyWords, maxResultsPerKeyword, maxDepth, 0, null);
-    }
-
-    public static SearchResults search(String token,
-                                       byte[] label,
-                                       Set<Keyword> keyWords,
-                                       int maxResultsPerKeyword,
-                                       int maxDepth,
-                                       int insecureFetchChainsBatchSize,
                                        String baseUrl)
         throws CloudproofException {
         //
@@ -100,9 +126,6 @@ public final class FindexCloud extends FindexBase {
                 token,
                 labelPointer, label.length,
                 wordsJson,
-                maxResultsPerKeyword,
-                maxDepth,
-                insecureFetchChainsBatchSize,
                 baseUrl);
             if (ffiCode != 0) {
                 // Retry with correct allocated size
@@ -112,9 +135,6 @@ public final class FindexCloud extends FindexBase {
                     token,
                     labelPointer, label.length,
                     wordsJson,
-                    maxResultsPerKeyword,
-                    maxDepth,
-                    insecureFetchChainsBatchSize,
                     baseUrl);
                 if (ffiCode != 0) {
                     throw new CloudproofException(get_last_error(4095));
@@ -130,7 +150,7 @@ public final class FindexCloud extends FindexBase {
     static public class SearchRequest extends FindexBase.SearchRequest<SearchRequest> {
         private String token;
 
-        private String baseUrl;
+        private String baseUrl = System.getenv("COSMIAN_FINDEX_CLOUD_BASE_URL");
 
         public SearchRequest(String token, byte[] label) {
             this.token = token;
@@ -161,7 +181,7 @@ public final class FindexCloud extends FindexBase {
     static public class IndexRequest extends FindexBase.IndexRequest<IndexRequest> {
         private String token;
 
-        private String baseUrl;
+        private String baseUrl = System.getenv("COSMIAN_FINDEX_CLOUD_BASE_URL");
 
         public IndexRequest(String token, byte[] label) {
             this.token = token;