+++ src/main/java/com/cosmian/jna/findex/ffi/FindexUserCallbacks.java
@@ -4,9 +4,11 @@ import java.util.List;
 import java.util.Map;
 import java.util.Set;
 
+import com.cosmian.jna.findex.serde.Tuple;
 import com.cosmian.jna.findex.structs.ChainTableValue;
 import com.cosmian.jna.findex.structs.EntryTableValue;
 import com.cosmian.jna.findex.structs.EntryTableValues;
+import com.cosmian.jna.findex.structs.IndexedValue;
 import com.cosmian.jna.findex.structs.Location;
 import com.cosmian.jna.findex.structs.Uid32;
 import com.cosmian.utils.CloudproofException;
@@ -19,7 +21,7 @@ public interface FindexUserCallbacks {
     }
 
     interface DBFetchEntry {
-        public Map<Uid32, EntryTableValue> fetch(List<Uid32> uids) throws CloudproofException;
+        public List<Tuple<Uid32, EntryTableValue>> fetch(List<Uid32> uids) throws CloudproofException;
     }
 
     interface DBFetchChain {