+++ src/main/java/com/cosmian/jna/findex/ffi/UpdateLines.java
@@ -3,6 +3,7 @@ package com.cosmian.jna.findex.ffi;
 import java.util.List;
 import java.util.Map;
 
+import com.cosmian.jna.findex.FindexCallbackException;
 import com.cosmian.jna.findex.ffi.FindexNativeWrapper.UpdateLinesCallback;
 import com.cosmian.jna.findex.ffi.FindexUserCallbacks.DBUpdateLines;
 import com.cosmian.jna.findex.serde.Leb128Reader;
@@ -28,23 +29,26 @@ public class UpdateLines implements UpdateLinesCallback {
                      Pointer newChainsPointer,
                      int newChainsLength)
         throws CloudproofException {
-
-        byte[] removedChainsBytes = new byte[removedChainsLength];
-        removedChainsPointer.read(0, removedChainsBytes, 0, removedChainsLength);
-        List<Uid32> removedChains = Leb128Reader.deserializeCollection(Uid32.class, removedChainsBytes);
-
-        byte[] newEntriesBytes = new byte[newEntriesLength];
-        newEntriesPointer.read(0, newEntriesBytes, 0, newEntriesLength);
-        Map<Uid32, EntryTableValue> newEntries =
-            Leb128Reader.deserializeMap(Uid32.class, EntryTableValue.class, newEntriesBytes);
-
-        byte[] newChainsBytes = new byte[newChainsLength];
-        newChainsPointer.read(0, newChainsBytes, 0, newChainsLength);
-        Map<Uid32, ChainTableValue> newChains =
-            Leb128Reader.deserializeMap(Uid32.class, ChainTableValue.class, newChainsBytes);
-
-        this.update.update(removedChains, newEntries, newChains);
-        return 0;
+        try {
+            byte[] removedChainsBytes = new byte[removedChainsLength];
+            removedChainsPointer.read(0, removedChainsBytes, 0, removedChainsLength);
+            List<Uid32> removedChains = Leb128Reader.deserializeCollection(Uid32.class, removedChainsBytes);
+    
+            byte[] newEntriesBytes = new byte[newEntriesLength];
+            newEntriesPointer.read(0, newEntriesBytes, 0, newEntriesLength);
+            Map<Uid32, EntryTableValue> newEntries =
+                Leb128Reader.deserializeMap(Uid32.class, EntryTableValue.class, newEntriesBytes);
+    
+            byte[] newChainsBytes = new byte[newChainsLength];
+            newChainsPointer.read(0, newChainsBytes, 0, newChainsLength);
+            Map<Uid32, ChainTableValue> newChains =
+                Leb128Reader.deserializeMap(Uid32.class, ChainTableValue.class, newChainsBytes);
+    
+            this.update.update(removedChains, newEntries, newChains);
+            return 0;
+        } catch (CloudproofException e) {
+            return FindexCallbackException.record(e);
+        }
     }
 
 }