+++ src/test/java/com/cosmian/findex/Sqlite.java
@@ -20,6 +20,7 @@ import java.util.stream.Collectors;
 
 import com.cosmian.jna.findex.Database;
 import com.cosmian.jna.findex.serde.Leb128ByteArray;
+import com.cosmian.jna.findex.serde.Tuple;
 import com.cosmian.jna.findex.structs.ChainTableValue;
 import com.cosmian.jna.findex.structs.EntryTableValue;
 import com.cosmian.jna.findex.structs.EntryTableValues;
@@ -108,7 +109,7 @@ public class Sqlite extends Database implements Closeable {
         return uids;
     }
 
-    public Map<Uid32, EntryTableValue> fetchEntryTableItems(List<Uid32> uids) throws SQLException {
+    public List<Tuple<Uid32, EntryTableValue>> fetchEntryTableItems(List<Uid32> uids) throws SQLException {
         PreparedStatement pstmt = this.connection
             .prepareStatement(
                 "SELECT uid, value FROM entry_table WHERE uid IN (" + questionMarks(uids.size()) + ")");
@@ -123,11 +124,11 @@ public class Sqlite extends Database implements Closeable {
         //
         // Recover all results
         //
-        HashMap<Uid32, EntryTableValue> uidsAndValues = new HashMap<>(uids.size(), 1);
+        ArrayList<Tuple<Uid32, EntryTableValue>> uidsAndValues = new ArrayList<>();
         while (rs.next()) {
-            uidsAndValues.put(
+            uidsAndValues.add(new Tuple<>(
                 new Uid32(rs.getBytes("uid")),
-                new EntryTableValue(rs.getBytes("value")));
+                new EntryTableValue(rs.getBytes("value"))));
         }
         return uidsAndValues;
     }
@@ -281,7 +282,7 @@ public class Sqlite extends Database implements Closeable {
     }
 
     @Override
-    protected Map<Uid32, EntryTableValue> fetchEntries(List<Uid32> uids) throws CloudproofException {
+    protected List<Tuple<Uid32, EntryTableValue>> fetchEntries(List<Uid32> uids) throws CloudproofException {
         try {
             return fetchEntryTableItems(uids);
         } catch (SQLException e) {